# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MyAgenda API
  version: 1.0.0
  description: API para gestión de usuarios, compañías y contactos (clientes) con autenticación JWT.
servers:
  - url: http://localhost:5000
tags:
  - name: Autenticación
    description: Endpoints de autenticación JWT
  - name: Usuarios
    description: Gestión de usuarios
  - name: Compañías
    description: Gestión de compañías
  - name: Clientes
    description: Gestión de clientes
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          description: Mensaje o lista de mensajes de error
          oneOf:
            - type: string
            - type: array
              items:
                type: string
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Email o nickname
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 1800
    UserCreate:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        nickname:
          type: string
          nullable: true
          maxLength: 50
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
        company:
          type: string
          nullable: true
          maxLength: 100
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
    UserRead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        nickname:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CompanyCreate:
      type: object
      required: [nit, business_name]
      properties:
        nit:
          type: string
          description: NIT colombiano (9 dígitos base + DV, e.g., 800197268-4)
        business_name:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          default: ACTIVE
    Company:
      allOf:
        - $ref: '#/components/schemas/CompanyCreate'
        - type: object
          properties:
            id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    ClientCreate:
      type: object
      required: [company_id]
      properties:
        company_id:
          type: string
        contact_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          default: ACTIVE
    Client:
      allOf:
        - $ref: '#/components/schemas/ClientCreate'
        - type: object
          properties:
            id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
paths:
  /login:
    post:
      security: []
      tags: [Autenticación]
      summary: Inicio de sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Petición inválida (body mal formado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth:
    get:
      tags: [Autenticación]
      summary: Validar token y obtener usuario autenticado
      responses:
        '200':
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          description: Token ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    post:
      security: []
      tags: [Usuarios]
      summary: Crear usuario (público)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email o nickname en uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}:
    patch:
      tags: [Usuarios]
      summary: Actualizar usuario (protegido)
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Nickname en uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/activate:
    post:
      tags: [Usuarios]
      summary: Activar usuario (protegido)
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario activado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/deactivate:
    post:
      tags: [Usuarios]
      summary: Desactivar usuario (protegido)
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario desactivado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /companies:
    get:
      tags: [Compañías]
      summary: Listar compañías (protegido)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Resultado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Compañías]
      summary: Crear compañía (protegido)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
      responses:
        '201':
          description: Compañía creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: NIT duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /companies/{company_id}:
    get:
      tags: [Compañías]
      summary: Obtener compañía por id (protegido)
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compañía
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Compañías]
      summary: Actualizar compañía (protegido)
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
      responses:
        '200':
          description: Compañía actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: NIT duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Compañías]
      summary: Eliminar compañía (protegido)
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminada
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clients:
    get:
      tags: [Clientes]
      summary: Listar clientes (protegido)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: include_company
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Resultado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Client'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Clientes]
      summary: Crear cliente (protegido)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email o teléfono duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clients/{client_id}:
    put:
      tags: [Clientes]
      summary: Actualizar cliente (protegido)
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email o teléfono duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Clientes]
      summary: Eliminar cliente (protegido)
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminado
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clients/{client_id}/deactivate:
    post:
      tags: [Clientes]
      summary: Inactivar cliente (protegido)
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente inactivado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  